requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
openai>=1.0
PyPDF2>=3.0
python-docx>=1.0
requests>=2.31
google-api-python-client>=2.0
google-auth>=2.0
